Moved the core parameters from the wx gui to ConfigManager
git-svn-id: https://svn.code.sf.net/p/vbam/code/trunk@1363 a31d4220-a93d-0410-bf67-fe4944624d44

+++ src/common/ConfigManager.h
@@ -136,7 +136,7 @@ extern int winPauseNextFrame;
 extern u32 autoFrameSkipLastTime;
 extern u32 movieLastJoypad;
 extern u32 movieNextJoypad;
-extern unsigned short throttle;
+extern int throttle;
 
 #define MAX_CHEATS 100
 extern int preparedCheats;
